+++ lib/src/effects/color_transition_effect.dart
@@ -1,17 +1,17 @@
 import 'package:flutter/material.dart';
+
 import '../painters/color_transition_painter.dart';
 import '../painters/indicator_painter.dart';
-
 import 'indicator_effect.dart';
 
 /// Holds [TransitionPainter]
-class ColorTransitionEffect extends IndicatorEffect {
+class ColorTransitionEffect extends BasicIndicatorEffect {
   // The active dot strokeWidth
   final double activeStrokeWidth;
 
   const ColorTransitionEffect({
     this.activeStrokeWidth = 1.5,
-    double offset,
+    double offset = 16.0,
     double dotWidth = 16.0,
     double dotHeight = 16.0,
     double spacing = 8.0,
@@ -20,9 +20,7 @@ class ColorTransitionEffect extends IndicatorEffect {
     Color activeDotColor = Colors.indigo,
     double strokeWidth = 1.0,
     PaintingStyle paintStyle = PaintingStyle.fill,
-  })  : assert(activeDotColor != null),
-        assert(activeStrokeWidth != null),
-        super(
+  }) : super(
             dotWidth: dotWidth,
             dotHeight: dotHeight,
             spacing: spacing,
@@ -34,6 +32,10 @@ class ColorTransitionEffect extends IndicatorEffect {
 
   @override
   IndicatorPainter buildPainter(int count, double offset) {
-    return TransitionPainter(count: count, offset: offset, effect: this);
+    return TransitionPainter(
+      count: count,
+      offset: offset,
+      effect: this,
+    );
   }
 }