+++ lib/src/effects/jumping_dot_effect.dart
@@ -1,17 +1,23 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:smooth_page_indicator/src/painters/indicator_painter.dart';
 import 'package:smooth_page_indicator/src/painters/jumping_dot_painter.dart';
 
 import 'indicator_effect.dart';
 
-class JumpingDotEffect extends IndicatorEffect {
-  // Defines how high the dot will jump
-  final double elevation;
+class JumpingDotEffect extends BasicIndicatorEffect {
+  // The maximum scale the dot will hit while jumping
+  final double jumpScale;
+
+  // The vertical offset of the jumping dot
+  final double verticalOffset;
 
   const JumpingDotEffect({
     Color activeDotColor = Colors.indigo,
-    this.elevation = 15.0,
-    double offset,
+    this.jumpScale = 1.4,
+    this.verticalOffset = 0.0,
+    double offset = 16.0,
     double dotWidth = 16.0,
     double dotHeight = 16.0,
     double spacing = 8.0,
@@ -19,9 +25,7 @@ class JumpingDotEffect extends IndicatorEffect {
     Color dotColor = Colors.grey,
     double strokeWidth = 1.0,
     PaintingStyle paintStyle = PaintingStyle.fill,
-  })  : assert(activeDotColor != null),
-        assert(elevation != null),
-        super(
+  }) : super(
             dotWidth: dotWidth,
             dotHeight: dotHeight,
             spacing: spacing,
@@ -30,6 +34,14 @@ class JumpingDotEffect extends IndicatorEffect {
             paintStyle: paintStyle,
             dotColor: dotColor,
             activeDotColor: activeDotColor);
+
+  @override
+  Size calculateSize(int count) {
+    return Size(
+      dotWidth * count + (spacing * (count - 1)),
+      max(dotHeight, dotHeight * jumpScale) + verticalOffset.abs(),
+    );
+  }
 
   @override
   IndicatorPainter buildPainter(int count, double offset) {