+++ lib/src/effects/swap_effect.dart
@@ -4,19 +4,20 @@ import 'package:smooth_page_indicator/src/painters/swap_painter.dart';
 
 import 'indicator_effect.dart';
 
-class SwapEffect extends IndicatorEffect {
+class SwapEffect extends BasicIndicatorEffect {
+  final SwapType type;
   const SwapEffect({
     Color activeDotColor = Colors.indigo,
-    double offset,
+    double offset = 16.0,
     double dotWidth = 16.0,
     double dotHeight = 16.0,
     double spacing = 8.0,
     double radius = 16,
     Color dotColor = Colors.grey,
     double strokeWidth = 1.0,
+    this.type = SwapType.normal,
     PaintingStyle paintStyle = PaintingStyle.fill,
-  })  : assert(activeDotColor != null),
-        super(
+  }) : super(
           dotWidth: dotWidth,
           dotHeight: dotHeight,
           spacing: spacing,
@@ -28,7 +29,24 @@ class SwapEffect extends IndicatorEffect {
         );
 
   @override
+  Size calculateSize(int count) {
+    var height = dotHeight;
+    if (type == SwapType.zRotation) {
+      height += height * .2;
+    } else if (type == SwapType.yRotation) {
+      height += dotWidth + spacing;
+    }
+    return Size(dotWidth * count + (spacing * count), height);
+  }
+
+  @override
   IndicatorPainter buildPainter(int count, double offset) {
+    assert(
+      offset.ceil() < count,
+      'SwapEffect does not support infinite looping.',
+    );
     return SwapPainter(count: count, offset: offset, effect: this);
   }
 }
+
+enum SwapType { normal, yRotation, zRotation }