+++ lib/src/painters/indicator_painter.dart
@@ -1,17 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:smooth_page_indicator/src/effects/indicator_effect.dart';
 
-abstract class IndicatorPainter extends CustomPainter {
-  /// The raw offset from the [PageController].page
-  /// This holds the directional offset
-  final double offset;
-
+abstract class BasicIndicatorPainter extends IndicatorPainter {
   /// The count of pages
   final int count;
 
   // The provided effect is passed to this super class
   // to make some calculations and paint still dots
-  final IndicatorEffect _effect;
+  final BasicIndicatorEffect _effect;
 
   /// Inactive dot paint or base paint in one-color effects.
   final Paint dotPaint;
@@ -19,16 +15,16 @@ abstract class IndicatorPainter extends CustomPainter {
   /// The Radius of all dots
   final Radius dotRadius;
 
-  IndicatorPainter(
-    this.offset,
+  BasicIndicatorPainter(
+    double offset,
     this.count,
     this._effect,
-  )   : assert(offset.ceil() < count, 'Current page is out of bounds'),
-        dotRadius = Radius.circular(_effect.radius),
+  )   : dotRadius = Radius.circular(_effect.radius),
         dotPaint = Paint()
           ..color = _effect.dotColor
           ..style = _effect.paintStyle
-          ..strokeWidth = _effect.strokeWidth;
+          ..strokeWidth = _effect.strokeWidth,
+        super(offset);
 
   // The distance between dot lefts
   double get distance => _effect.dotWidth + _effect.spacing;
@@ -37,16 +33,41 @@ abstract class IndicatorPainter extends CustomPainter {
     for (var i = 0; i < count; i++) {
       final xPos = (i * distance);
       final yPos = size.height / 2;
-      final bounds = Rect.fromLTRB(xPos, yPos - _effect.dotHeight / 2,
-          xPos + _effect.dotWidth, yPos + _effect.dotHeight / 2);
+      final bounds = Rect.fromLTRB(
+        xPos,
+        yPos - _effect.dotHeight / 2,
+        xPos + _effect.dotWidth,
+        yPos + _effect.dotHeight / 2,
+      );
       var rect = RRect.fromRectAndRadius(bounds, dotRadius);
       canvas.drawRRect(rect, dotPaint);
     }
   }
 
+  RRect calcPortalTravel(Size size, double offset, double dotOffset) {
+    final yPos = size.height / 2;
+    var width = dotOffset * (_effect.dotHeight / 2);
+    var height = dotOffset * (_effect.dotWidth / 2);
+    var xPos = offset;
+    return RRect.fromLTRBR(
+      xPos - height,
+      yPos - width,
+      xPos + height,
+      yPos + width,
+      Radius.circular(dotRadius.x * dotOffset),
+    );
+  }
+}
+
+abstract class IndicatorPainter extends CustomPainter {
+  /// The raw offset from the [PageController].page
+  final double offset;
+
+  const IndicatorPainter(this.offset);
+
   @override
   bool shouldRepaint(IndicatorPainter oldDelegate) {
-    // only repaint if the raw offset changes
+    // only repaint if the offset changes
     return oldDelegate.offset != offset;
   }
 }