+++ lib/src/painters/slide_painter.dart
@@ -3,13 +3,13 @@ import 'package:smooth_page_indicator/src/effects/slide_effect.dart';
 
 import 'indicator_painter.dart';
 
-class SlidePainter extends IndicatorPainter {
+class SlidePainter extends BasicIndicatorPainter {
   final SlideEffect effect;
 
   SlidePainter({
-    @required this.effect,
-    @required int count,
-    @required double offset,
+    required this.effect,
+    required int count,
+    required double offset,
   }) : super(offset, count, effect);
 
   @override
@@ -18,8 +18,23 @@ class SlidePainter extends IndicatorPainter {
 
     paintStillDots(canvas, size);
 
-    final activePaint = Paint()..color = effect.activeDotColor;
-    final xPos = (offset * distance);
+    final activeDotPainter = Paint()..color = effect.activeDotColor;
+    final dotOffset = offset - offset.toInt();
+    // handle dot travel from end to start (for infinite pager support)
+    if (offset > count - 1) {
+      final startDot = calcPortalTravel(size, effect.dotWidth / 2, dotOffset);
+      canvas.drawRRect(startDot, activeDotPainter);
+
+      final endDot = calcPortalTravel(
+        size,
+        ((count - 1) * distance) + (effect.dotWidth / 2),
+        1 - dotOffset,
+      );
+      canvas.drawRRect(endDot, activeDotPainter);
+      return;
+    }
+
+    final xPos = offset * distance;
     final yPos = size.height / 2;
     final rRect = RRect.fromLTRBR(
       xPos,
@@ -29,6 +44,6 @@ class SlidePainter extends IndicatorPainter {
       dotRadius,
     );
 
-    canvas.drawRRect(rRect, activePaint);
+    canvas.drawRRect(rRect, activeDotPainter);
   }
 }